Moving marker clustering helper functions to designated file
